Reorganize config module
